training/sdxl/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
xformers>=0.0.22
pillow>=10.0.0
numpy>=1.24.0
wandb>=0.15.0
datasets>=2.14.0
peft>=0.6.0

